Fix Robot story
